# openapi.yaml
openapi: 3.0.3
info:
  title: Babbel API
  description: |
    HTTP API for generating audio news bulletins. Combines news stories with station-specific jingles using FFmpeg.
    
    ## Authentication
    
    - Local: Username/password with session cookies
    - OAuth/OIDC: Azure AD, Google, Okta support
    - Account lockout protection
    
    ## Authorization
    
    Role-based access control:
    - admin: Full access including user management
    - editor: Content management (stations, voices, stories, bulletins)
    - viewer: Read-only access
    
    ## Architecture
    
    Station-voice junction table:
    - Each voice has different jingles per station
    - Mix point defines when voice starts over jingle
    - Files stored as `station_{id}_voice_{id}_jingle.wav`
  version: 1.0.0
  x-changelog:
    1.0.0:
      - "Initial stable release"
      - "HTTP API for audio news bulletin generation"
      - "Station-specific voice jingles with mix points"
      - "Local and OAuth/OIDC authentication"
      - "Role-based access control (admin, editor, viewer)"
      - "Soft delete for stories and users"
      - "FFmpeg audio processing"
      - "Pagination and filtering on all list endpoints"
      - "RESTful design for radio automation"
      - "Removed redundant /broadcasts endpoint"
  contact:
    name: Streekomroep ZuidWest
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://your-api-server.com/api/v1
...
security:
  - cookieAuth: []
paths:
...
  /session/login:
    post:
      summary: Login with username and password (local authentication)
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                description: Session cookie
        '401':
          $ref: '#/components/responses/Unauthorized'
  /session/oauth/start:
    get:
      summary: Start OAuth/OIDC authentication flow
      tags:
        - Authentication
      security: []
      description: Redirects to the configured OAuth provider (Azure AD, Google, etc.)
      responses:
        '302':
          description: Redirect to OAuth provider
          headers:
            Location:
              schema:
                type: string
                description: OAuth provider authorization URL
        '400':
          description: OAuth authentication is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /session/oauth/callback:
    get:
      summary: OAuth callback endpoint
      tags:
        - Authentication
      security: []
      description: Handles the OAuth callback from the provider
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from OAuth provider
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: CSRF protection state
      responses:
        '302':
          description: Redirect to dashboard after successful authentication
          headers:
            Location:
              schema:
                type: string
                default: /dashboard
        '400':
          $ref: '#/components/responses/BadRequest'
  /session:
    get:
      summary: Get current user information
      tags:
        - Authentication
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  role:
                    type: string
                    enum: [admin, editor, viewer]
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Logout and destroy session
      tags:
        - Authentication
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
  /stations:
    get:
      summary: List all stations
      tags:
        - Stations
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of stations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Station'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
    post:
      summary: Create a new station
      tags:
        - Stations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationInput'
      responses:
        '201':
          description: Station created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '400':
          $ref: '#/components/responses/BadRequest'
  /stations/{id}:
    get:
      summary: Get station by ID
      tags:
        - Stations
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Station details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update station
      tags:
        - Stations
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationInput'
      responses:
        '200':
          description: Station updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete station
      tags:
        - Stations
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Station deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /voices:
    get:
      summary: List all voices
      tags:
        - Voices
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of voices
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Voice'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
    post:
      summary: Create a new voice
      tags:
        - Voices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Voice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voice'
        '400':
          $ref: '#/components/responses/BadRequest'
  /voices/{id}:
    get:
      summary: Get voice by ID
      tags:
        - Voices
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Voice details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voice'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update voice
      tags:
        - Voices
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Voice updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voice'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete voice
      tags:
        - Voices
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Voice deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /stories:
    get:
      summary: List all stories
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: include_deleted
          in: query
          schema:
            type: boolean
            default: false
          description: Include soft-deleted stories in the results
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, expired]
          description: Filter by status
        - name: voice_id
          in: query
          schema:
            type: integer
          description: Filter by voice ID
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter stories active on this date
        - name: weekday
          in: query
          schema:
            type: string
            enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
          description: Filter by weekday
      responses:
        '200':
          description: List of stories
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Story'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
    post:
      summary: Create a new story
      tags:
        - Stories
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StoryInput'
      responses:
        '201':
          description: Story created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '400':
          $ref: '#/components/responses/BadRequest'
  /stories/{id}/audio:
    get:
      summary: Download story audio file
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Audio file
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '404':
          description: Story not found or no audio file
  /stories/{id}:
    get:
      summary: Get story by ID
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Story details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update story
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StoryInput'
      responses:
        '200':
          description: Story updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete story (soft delete)
      description: |
        Performs a soft delete by setting deleted_at timestamp. 
        The story will no longer appear in listings unless include_deleted=true is specified.
        Story data and audio files are preserved for potential restoration.
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Story soft deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update story state
      description: Update story status (draft, active, expired) or other state fields without requiring full story data.
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [draft, active, expired]
                  description: Story status
              example:
                status: "active"
      responses:
        '200':
          description: Story state updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Story status updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /stories/{id}/bulletins:
    get:
      summary: Get bulletin history for a story
      description: Returns all bulletins that have included this specific story, ordered by most recent first.
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Story bulletin history
          content:
            application/json:
              schema:
                type: object
                properties:
                  story_id:
                    type: integer
                  story_title:
                    type: string
                  bulletins:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        station_id:
                          type: integer
                        station_name:
                          type: string
                        filename:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        duration:
                          type: number
                        story_order:
                          type: integer
                        included_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
  /users:
    get:
      summary: List all users
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: include_suspended
          in: query
          schema:
            type: boolean
            default: false
          description: Include suspended users in the results
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, editor, viewer]
          description: Filter by user role
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
    post:
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /users/{id}:
    get:
      summary: Get user by ID
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update user
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Permanently delete user
      description: |
        Permanently deletes a user account and all associated data.
        This action cannot be undone. All active sessions for the user will be terminated.
        Cannot delete the last admin user.
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: User permanently deleted
        '400':
          description: Cannot delete the last admin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update user state
      description: Update user state (suspend/restore) without requiring full user data. This endpoint handles suspension and restoration of user accounts.
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [suspend, restore]
                  description: Action to perform on user account
              required:
                - action
              example:
                action: "suspend"
      responses:
        '200':
          description: User state updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User suspended successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/password:
    put:
      summary: Change user password
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  minLength: 8
                  description: New password (minimum 8 characters)
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /bulletins:
    get:
      summary: List bulletins
      description: |
        Returns a paginated list of generated bulletins with optional filtering.
        
        Query parameters:
        - `station_id` - Filter bulletins by station ID
        - `limit` - Maximum number of bulletins to return (default: 20)
        - `offset` - Number of bulletins to skip for pagination (default: 0)
      tags:
        - Bulletin
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: station_id
          in: query
          required: false
          schema:
            type: integer
          description: Filter by station ID
        - name: include_stories
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Include detailed story information in bulletin response
      responses:
        '200':
          description: List of bulletins
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BulletinResponse'
                  total:
                    type: integer
                    description: Total number of bulletins
                  limit:
                    type: integer
                    description: Number of items per page
                  offset:
                    type: integer
                    description: Number of items skipped
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /stations/{id}/bulletins/latest:
    get:
      summary: Get latest bulletin for a station
      description: Returns the most recently generated bulletin for a specific station. This endpoint is ideal for radio automation systems that need to fetch the latest news bulletin.
      tags:
        - Bulletin
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Latest bulletin found
          content:
            application/json:
              schema:
                type: object
                properties:
                  station_id:
                    type: integer
                    description: Station ID
                  station_name:
                    type: string
                    description: Station name
                  bulletin_url:
                    type: string
                    description: URL to download the bulletin
                  filename:
                    type: string
                    description: Bulletin filename
                  created_at:
                    type: string
                    format: date-time
                    description: When the bulletin was created
                  duration:
                    type: number
                    format: float
                    description: Duration in seconds
                  file_size:
                    type: integer
                    format: int64
                    description: File size in bytes
                  story_count:
                    type: integer
                    description: Number of stories in the bulletin
        '404':
          description: Station not found or no bulletin available
  /stations/{id}/bulletins/latest/audio:
    get:
      summary: Download latest bulletin audio for a station
      description: |
        Downloads the audio file of the most recently generated bulletin for a specific station.
        This endpoint is specifically designed for radio automation systems that need to directly
        fetch and play the latest news bulletin without additional API calls.
        
        Perfect for radio automation workflows where you need:
        - One URL call that always returns audio
        - Latest bulletin audio for immediate playback
        - No metadata parsing required
      tags:
        - Bulletin
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Latest bulletin audio file
          content:
            audio/wav:
              schema:
                type: string
                format: binary
                description: WAV audio file of the latest bulletin
          headers:
            Content-Disposition:
              schema:
                type: string
                description: Attachment with filename
                example: 'attachment; filename="bulletin_1_20240315_143022.wav"'
            Content-Length:
              schema:
                type: integer
                description: File size in bytes
        '404':
          description: Station not found or no bulletin available for this station
  /bulletins/{id}/audio:
    get:
      summary: Download bulletin audio file
      tags:
        - Bulletin
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Audio file
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '404':
          description: Bulletin not found or no audio file
...
  /station_voices/{id}/audio:
    get:
      summary: Download station-voice jingle file
      tags:
        - Station-Voices
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Jingle audio file
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '404':
          description: Station-voice not found or no jingle file
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: babbel_session
      description: Session-based authentication using secure cookies
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...